Move firebase to utils folder
